pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corex_eval"
version = "0.1.0"
description = "Shared evaluation library for the CoREx LLM benchmarking project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Required for semantic similarity in annotation evaluation
embeddings = [
    "torch",
    "sentence-transformers>=2.2",
]
# Required for development and running tests
dev = [
    "pytest>=7.4",
    "pytest-cov",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["corex_eval*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]